+++ Zephyr/Options.lua
@@ -1,11 +1,22 @@
 local addonName, ns = ...
-local addonTitle = GetAddOnInfo(addonName)
+local addonTitle = GetAddOnInfo(addonName, "addonName")
 local exclamation = "|TInterface\\Cursor\\Quest:12:12:1:0:64:64:9:60:9:60|t"
 
 ZPR.defaults = {
 	profile = {
 		modules = {
+			-- General
+
+			-- Action Bars
+
+			-- Unit Frames
+			pvpIcon = true,
+			feedbackText = true,
 			classColors = true,
+			combatIndicator = true,
+
+			-- Miscellaneous
+
 		}
 	}
 }
@@ -58,23 +69,71 @@ ZPR.options = {
 					type = "group",
 					name = "General",
 					desc = "Configure options for the General module.",
+					order = 1,
 					args = {}
 				},
 				actionBars = {
 					type = "group",
 					name = "Action Bars",
 					desc = "Configure options for the Action Bars module.",
+					order = 2,
 					args = {}
 				},
 				unitFrames = {
 					type = "group",
 					name = "Unit Frames",
 					desc = "Configure options for the Unit Frames module.",
+					order = 3,
 					args = {
+						header_visibility = {
+							type = "header",
+							name = "Visibility",
+							width = "full",
+							order = 1
+						},
+						spacer3 = {
+							name = "",
+							type = "description",
+							order = 2
+						},
+						pvpIcon = {
+							type = "toggle",
+							name = "PvP Icon",
+							desc = "Enable or disable the PvP icon for unit frames.",
+							width = "half",
+							order = 3,
+							set = function(info, value)
+								ZPR.db.profile.modules.pvpIcon = value
+								ZPR:ToggleModules()
+							end,
+							get = function(info)
+								return ZPR.db.profile.modules.pvpIcon
+							end
+						},
+						feedbackText = {
+							type = "toggle",
+							name = "Feedback Text",
+							desc = "Disable or enable feedback text for the Player & Pet Portraits.",
+							order = 4,
+							set = function(info, value)
+								ZPR.db.profile.modules.feedbackText = value
+								ZPR:ToggleModules()
+							end,
+							get = function(info)
+								return ZPR.db.profile.modules.feedbackText
+							end
+						},
+						header_General = {
+							type = "header",
+							name = "General Unit Frame Customization",
+							width = "full",
+							order = 5
+						},
 						classColors = {
 							type = "toggle",
 							name = "Class Colors",
 							desc = "Enable or disable class colors for unit frames.",
+							order = 6,
 							set = function(info, value)
 								ZPR.db.profile.modules.classColors = value
 								ZPR:ToggleModules()
@@ -82,13 +141,48 @@ ZPR.options = {
 							get = function(info)
 								return ZPR.db.profile.modules.classColors
 							end
-						}
-					}
+						},
+						repColor = {
+							type = "toggle",
+							name = "Rep Colors",
+							desc = "Disable or enable reputation colors for unit frames.",
+							width = "half",
+							order = 7,
+							set = function(info, value)
+								ZPR.db.profile.modules.repColor = value
+								ZPR:ToggleModules()
+							end,
+							get = function(info)
+								return ZPR.db.profile.modules.repColor
+							end
+						},
+						header_Misc = {
+							type = "header",
+							name = "Miscellaneous Unit Frame Customization",
+							width = "full",
+							order = 8
+
+						},
+						combatIndicator = {
+							type = "toggle",
+							name = "Combat Indicator",
+							desc = "Enable or disable a icon to the right of target/focus frames to indicate combat status.",
+							order = 9,
+							set = function(info, value)
+								ZPR.db.profile.modules.combatIndicator = value
+								ZPR:ToggleModules()
+							end,
+							get = function(info)
+								return ZPR.db.profile.modules.combatIndicator
+							end
+						},
+					},
 				},
 				misc = {
 					type = "group",
 					name = "Misc",
 					desc = "Configure options for the Misc module.",
+					order = 5,
 					args = {}
 				},
 			}