+++ CHANGELOG.md
@@ -14,3 +14,10 @@ All notable changes to this project will be documented in this file.
 
 -   Fixed a formatting issue in the `Zephyr.toc` file.
 -   Updated the `README.md` file.
+
+## [1.0.3] 2022-12-21
+
+-   Implemented `Class Colors` for Unit Frames.
+-   Implemented `Hide Rep Color` for Unit Frames.
+-   Implemented `Hide Feedback Text` for Unit Frames.
+-   Implemented `Hide PvP Icon` for Unit Frames.